% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{dist.func}
\alias{dist.func}
\title{Calculating the distance between each row of the matrix and the centroid}
\usage{
dist.func(mndata, standardize = TRUE, method = "euclidean",
  centroid = NULL)
}
\arguments{
\item{mndata}{matrix dataset with numeric columns.}

\item{standardize}{is a booleon variable to denote whether columns should first be scaled to 0-1.}

\item{method}{dictates how distance is computed.  Currently, only "euclidean" is implemented.}

\item{centroid}{allows user to specify the centroid values for each column.  Otherwise, the mean of each column is taken by default.}
}
\description{
\code{dist.func} returns a sum total of all distances for each row of the matrix from the centroid.
}

