% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impmat.R
\name{impmat}
\alias{impmat}
\title{Imputation function}
\usage{
impmat(data)
}
\arguments{
\item{data}{dataset to be imputed.  Should contain all numeric values with rows=Genes and columns=Samples.}
}
\description{
\code{impmat} imputes a dataset entered as rows=Genes, columns=Samples.   NA values are imputed by row.  Currently, impvalue function imputes these as the max value of the row, 
the largest observed value for a given Gene across all Samples.
}

