% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreeningNA.R
\name{scr}
\alias{scr}
\title{Replacing values above ct to NA}
\usage{
scr(dat, threshold, missing.percent.thresh)
}
\arguments{
\item{dat}{the dataset with rows = Genes and columns = Samples.}

\item{threshold}{Ct value at or above which all values are set to NA.}

\item{missing.percent.thresh}{a number between 0 to 100 denoting the percentage of allowable missing.  If the percentage of missing is greater than this value, the row (gene) is not retained.}
}
\description{
\code{scr} screens the dataset in wide format according to Ct threshold and missing.percent.thresh (for each gene, the percentage of samples is missing).
sets all values above Ct threshold to NA and then removes all rows of genes with number of missing above missing.percent.thresh.
}

